- Added Spell and mudprogcommand files - Removed the nuget library check-ins as the .nuget package will grab them now on a build request - added realm.library.lua and unit tests to begin working on converting mud progs to lua progs - continued work on magic, misc, and mud progs

+++ trunk/SmaugCS/handler.cs
@@ -1130,7 +1130,7 @@ namespace SmaugCS
             return false;
         }
 
-        public static bool chance(CharacterInstance ch, short percent)
+        public static bool chance(CharacterInstance ch, int percent)
         {
             // TODO
             return false;
